Update Matroska implementation to support large files

+++ taglib/matroska/ebmlreader.cpp
@@ -4,7 +4,7 @@
 
 using namespace TagLib;
 
-Matroska::EBMLReader::EBMLReader(File *_file, ulong position)
+Matroska::EBMLReader::EBMLReader(File *_file, long long position)
   : file (_file),
     parent (0),
     offset (position),
@@ -15,7 +15,7 @@ Matroska::EBMLReader::EBMLReader(File *_file, ulong position)
   isSuccessRead = read();
 }
 
-Matroska::EBMLReader::EBMLReader(const EBMLReader &parent, ulong position, uint id, ulong size)
+Matroska::EBMLReader::EBMLReader(const EBMLReader &parent, long long position, uint id, long long size)
   : offset(position),
     ebmlId(id),
     dataOffset(position),
@@ -31,12 +31,12 @@ Matroska::MatroskaID Matroska::EBMLReader::id() const
   return static_cast<MatroskaID>(ebmlId);
 }
 
-ulong Matroska::EBMLReader::getDataSize() const
+long long Matroska::EBMLReader::getDataSize() const
 {
   return dataSize;
 }
 
-ulong Matroska::EBMLReader::size() const
+long long Matroska::EBMLReader::size() const
 {
   return (dataOffset - offset) + dataSize;
 }
@@ -51,7 +51,7 @@ String Matroska::EBMLReader::readString() const
   String result;
   if (!file || dataSize == 0) {
       return result;
-    }
+  }
   ByteVector data (readBytes());
   EBMLElement ebml (id(), data);
   result = ebml.getString();
@@ -64,9 +64,9 @@ ByteVector Matroska::EBMLReader::readBytes() const
 
   if (!file || dataSize == 0) {
       return result;
-    }
+  }
 
-  file->seek(static_cast<long>(dataOffset));
+  file->seek(dataOffset);
 
   result = file->readBlock(dataSize);
 
@@ -78,7 +78,7 @@ ulong Matroska::EBMLReader::readUInt() const
   ulong result = 0;
   if (!file || dataSize == 0) {
       return result;
-    }
+  }
   ByteVector data (readBytes());
   EBMLElement ebml (id(), data);
   return ebml.getUInt();
@@ -89,7 +89,7 @@ double Matroska::EBMLReader::readDouble() const
   double result = 0.0;
   if (!file || dataSize == 0) {
       return result;
-    }
+  }
   ByteVector data (readBytes());
   EBMLElement ebml (id(), data);
   return ebml.getDouble();
@@ -100,12 +100,12 @@ const Matroska::EBMLReader *Matroska::EBMLReader::getParent() const
   return parent;
 }
 
-ulong Matroska::EBMLReader::getOffset() const
+long long Matroska::EBMLReader::getOffset() const
 {
   return offset;
 }
 
-ulong Matroska::EBMLReader::getDataOffset() const
+long long Matroska::EBMLReader::getDataOffset() const
 {
   return dataOffset;
 }
@@ -114,22 +114,22 @@ bool Matroska::EBMLReader::read()
 {
   if (!isAbstract()) {
       return true;
-    }
+  }
 
   if (!file) {
       debug("EBMLReader - null file");
       return false;
-    }
+  }
 
-  if (offset >= static_cast<ulong>(file->length()) - 1) {
+  if (offset >= file->length() - 1) {
       debug("EBMLReader - invalid offset");
       return false;
-    }
+  }
 
   // Prepare for Consistency check
   uint ebmlIdCheck = ebmlId;
-  ulong ebmlSizeCheck = size();
-  file->seek(static_cast<long>(offset));
+  long long ebmlSizeCheck = size();
+  file->seek(offset);
 
   int idMaxLength = 4;
 
@@ -140,16 +140,16 @@ bool Matroska::EBMLReader::read()
   if (idLength > idMaxLength) {
       debug("Invalid EBML format read");
       return false;
-    }
+  }
 
   ByteVector idData = headerByteVector;
   if (idLength > 1) {
       idData.append(file->readBlock(static_cast<ulong>(idLength - 1)));
-    }
+  }
 
   ebmlId = idData.toUInt();
 
-  ByteVector sizeLengthVector = file->readBlock (1);
+  ByteVector sizeLengthVector = file->readBlock(1);
   int maxSizeLength = 8;
 
   mask = 0x80; // reset mask
@@ -157,30 +157,30 @@ bool Matroska::EBMLReader::read()
 
   if (sizeLength > 8) {
       sizeLength = 1; // Special: Empty element (all zero state)
-    } else {
+  } else {
       sizeLengthVector[0] &= static_cast<char>(mask - 1);  // Clear the marker bit
-    }
+  }
 
   // Now read the rest of the EBML element size
   if (sizeLength > 1) {
       sizeLengthVector.append(file->readBlock(static_cast<ulong>(sizeLength) - 1));
-    }
+  }
 
-  dataSize = static_cast<ulong>(sizeLengthVector.toLongLong());
+  dataSize = sizeLengthVector.toLongLong();
 
   // Special: Auto-size (0xFF byte)
   if (sizeLength == 1 && dataSize == 0x7F) {
       // Resolve auto-size to fill in to its containing element
-      ulong bound = parent ? parent->offset + parent->size() : static_cast<ulong>(file->length());
-      dataSize = bound - offset - static_cast<ulong>(idLength + sizeLength);
-    }
-  dataOffset = offset + static_cast<ulong>(idLength + sizeLength);
+      long long bound = parent ? parent->offset + parent->size() : file->length();
+      dataSize = bound - offset - idLength + sizeLength;
+  }
+  dataOffset = offset + idLength + sizeLength;
   // Consistency check: Detect descrepencies between read data and abstract data
   if ((ebmlIdCheck != 0 && ebmlIdCheck != ebmlId) ||
       (ebmlSizeCheck != 0 && ebmlSizeCheck != size())) {
       debug("Consistensy check failed");
       return false;
-    }
+  }
 
   return true;
 }
@@ -195,7 +195,7 @@ char Matroska::EBMLReader::getVINTLength(int maxSize, const ByteVector &data, in
   if (!data.size()) {
       debug("Invalid data for reading VINT");
       return -1;
-    }
+  }
   const char headerByte = data[0];
 
   char length = 1;
@@ -204,6 +204,6 @@ char Matroska::EBMLReader::getVINTLength(int maxSize, const ByteVector &data, in
   while (length <= maxSize && (headerByte & mask) == 0) {
       length++;
       mask >>= 1;
-    }
+  }
   return length;
 }
